src/microservices/events/requirements.txt
Flask==3.0.3
kafka-python==2.0.2